# DB_lab/openapi.yml
openapi: 3.0.0
info:
  title: DB Lab API
  version: 1.0.1

paths:

  /users:
    get:
      summary: Get all users
      description: Returns a list of all users in the database.
      responses:
        '200':
          description: Successful response with list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      summary: Create a new user
      description: Inserts a new user with name, email, and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully

  /users/{user_id}:
    get:
      summary: Get user by ID
      description: Returns a user by their ID.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    put:
      summary: Update user by ID
      description: Updates user attributes (name, email, password).
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully

    delete:
      summary: Delete user by ID
      description: Deletes a user from the system.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found


components:

  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: integer
      description: ID of the user

  schemas:

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string

    UserCreate:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string

    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string